backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
PyPDF2
python-dotenv 